requirements.txt
python-binance==1.0.19
pyTelegramBotAPI==4.18.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
python-dotenv==1.0.0
requests>=2.28.0
google-genai>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
importlib-metadata>=6.0.0
psycopg2-binary>=2.9.0
websockets>=12.0
